feat: refactor mobile login components and add hrms individual utility
